requirements.txt
# Use cpu version of torch
--extra-index-url https://download.pytorch.org/whl/cpu
--extra-index-url https://developer.download.nvidia.com/compute/redist

absl-py==1.3.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
dlio_profiler_py==0.0.3
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.7.0
google-pasta==0.2.0
grpcio==1.51.0
h5py==3.7.0
hydra-core==1.2.0
idna==3.4
keras==2.15.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.5
nvidia-dali-cuda110==1.34.0
oauthlib==3.2.2
omegaconf==2.2.3
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
protobuf==4.23.4
psutil==5.9.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.15.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.15.0
tensorflow-io==0.28.0
tensorflow-estimator==2.15.0
termcolor==2.1.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
typing_extensions==4.9.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
psutil>5.9.5
